Refactoring literallly everything
Well, because I was creating this game engine with no knowledge of
creating software as a 19-21 something year old. I've taken a good
look at my piece of software and honestly it is not good. To list off
a few examples
1. The way the code is not consistent at all, for two similar .cpp files
are coded different usually.
2. Files are not originizated properly.
3. Design patterns are not coded properly or consistently
4. Engine Systems are separated from DogEngine class, which is suppose
to be the head of all the game engine functionality.
5.comments are everywhere and somehow manage to be irrevlated and
  outdated.
6. way too much code in certain files, because it is just me working on
   this game engine, .cpp files need to be pretty manageable.
7. The scope of functionality of every source files is way off, usually too
   much. I need to be more consistent and interment the functionality of
   my source code.
8. Scope and planning of my game engine is out of hand. I need to
   properly restructure my work on my game engine.